+++ src/core/SparseFileWriter.h
@@ -1,14 +1,13 @@
 #pragma once
 
-#include <QMutex>
 #include <QString>
+#include <QMutex>
 #include <cstdint>
 
 namespace qtidm {
 
-class SparseFileWriter final {
+class SparseFileWriter {
 public:
-    SparseFileWriter() = default;
     ~SparseFileWriter();
 
     bool open(const QString& path, std::int64_t size, bool preserveUnknownLength = false);
@@ -21,12 +20,12 @@ private:
     bool map(std::int64_t offset, std::int64_t size);
     void unmap();
 
+    mutable QMutex mutex_;
     int fd_ = -1;
+    std::int64_t fileSize_ = 0;
     void* mapped_ = nullptr;
     std::int64_t mappedOffset_ = 0;
     std::int64_t mappedSize_ = 0;
-    std::int64_t fileSize_ = 0;
-    mutable QMutex mutex_;
     QString lastError_;
 };